requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
python-decouple
bcrypt
pyjwt